+++ routes/ClienteReservasRoutes.js
@@ -1,0 +1,10 @@
+// ClienteReservasRoutes.js
+const express = require('express');
+const router = express.Router();
+
+const {busquedaUser} = require('../controllers/ClienteReservasController')
+
+router.post('/busqueda', busquedaUser)
+
+module.exports = router;
+
