--- routes/cliente+reservasRoutes.js
@@ -1,10 +1,0 @@
-
-const express = require('express');
-const router = express.Router();
-const { getReservasCliente } = require('../controllers/cliente+reservasController'); 
-
-router.get('/reservas/cliente/:idCliente', async (req, res) => {
-  await getReservasCliente(req, res);
-});
-
-module.exports = router;
