+++ server.js
@@ -5,7 +5,6 @@ const routes = require('./routes/index')
 const cookieParser = require('cookie-parser')
 const cors = require('cors')
 const morgan = require('morgan')
-const isAdminMiddleware = require('./middlewares/isAdminMiddleware'); // Importar el middleware isAdminMiddleware
 
 require('dotenv').config()
 
@@ -28,9 +27,6 @@ app.use(morgan('dev'))
 
 //RUTAS
 app.use('/api', routes)
-
-// Middleware isAdmin para todas las rutas a partir de aquí
-app.use('/api', isAdminMiddleware);
 
 app.listen(port, () => {
   console.log('Server corriendo en el puerto: ' + port)
